Modularize Firebase auth & misc functions
Moved most Firebase auth functions to firebaseConfig.js to modularize and declutter for README 'How to Run' prep.
