+++ scripts/authDialogs.js
@@ -1,24 +1,8 @@
-// import { showNotification, updateUIAfterLogin } from "../scripts.js";    //was causing errors, leave commented out for now
-import { initializeApp } from "https://www.gstatic.com/firebasejs/11.3.1/firebase-app.js";
-import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword, signOut } from "https://www.gstatic.com/firebasejs/11.3.1/firebase-auth.js";
-
-// Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyCteFAmh1TjbbQB0hsbBwcbqwK8mofMO4Y",
-  authDomain: "b-coders-database.firebaseapp.com",
-  projectId: "b-coders-database",
-  storageBucket: "b-coders-database.appspot.com",
-  messagingSenderId: "268773123996",
-  appId: "1:268773123996:web:fec77ef63557a9c6b50a59",
-  measurementId: "G-92LTT20BXB"
-};
+import { auth, signInWithEmailAndPassword, createUserWithEmailAndPassword } from "./firebaseConfig.js";
 
 // ==========================
 // Authencation: Login and Logout + Dialogs
 // ==========================
-
-const app = initializeApp(firebaseConfig);
-const auth = getAuth(app);
 
 // Function to open the login dialog
 export function openLoginDialog() {
@@ -134,6 +118,67 @@ export function createSignupDialog() {
         closeDialogBoxes();
         }
     });
+}
+
+export function showNotification(message, color = "gold") {
+  // Remove existing notification if present
+  const existingNotification = document.getElementById("notification-box");
+  if (existingNotification) existingNotification.remove();
+
+  // Create notification container
+  const notification = document.createElement("div");
+  notification.id = "notification-box";
+  notification.classList.add("notification");
+  notification.style.backgroundColor = color;
+
+  // Create notification text
+  const messageText = document.createElement("span");
+  messageText.textContent = message;
+
+  // Create close button (X)
+  const closeButton = document.createElement("span");
+  closeButton.textContent = "✖";
+  closeButton.classList.add("close-button");
+  closeButton.addEventListener("click", () => fadeOutNotification(notification));
+
+  // Append elements to notification
+  notification.appendChild(messageText);
+  notification.appendChild(closeButton);
+  document.body.appendChild(notification);
+
+  // Timer to fade out and remove notification
+  let removeTimeout = setTimeout(() => {
+    fadeOutNotification(notification);
+  }, 3000); // 3 seconds
+
+  // Pause timer when hovering
+  notification.addEventListener("mouseenter", () => clearTimeout(removeTimeout));
+
+  // Resume timer when mouse leaves
+  notification.addEventListener("mouseleave", () => {
+    removeTimeout = setTimeout(() => fadeOutNotification(notification), 3000);
+  });
+}
+
+// Function to fade out notification before removing it
+function fadeOutNotification(notification) {
+  notification.classList.add("fade-out");
+  setTimeout(() => notification.remove(), 500); // Wait for fade-out transition
+}
+
+// Function to update the UI after user login/logout
+export function updateUIAfterLogin(user) {
+    const loginButton = document.getElementById("loginButton");
+  
+    if (user) {
+      loginButton.textContent = "Log Out";
+      loginButton.removeEventListener("click", openLoginDialog);
+      loginButton.addEventListener("click", logoutUser);
+    } else {
+      loginButton.textContent = "Log In";
+      loginButton.removeEventListener("click", logoutUser);
+      loginButton.addEventListener("click", openLoginDialog);
+    }
 }
 
 // Function to attempt user login
